Use InputStream instead of BufferedReader
Why?

Because "read" call on BufferedReader returns an integer which ranges
from 0-65535 which screws you if you are expecting return type to be a
character.
e.g I was expecting 0xB5 and I was getting 65533 and had no way to
convert it to B5 (not an easy one).
So I switched to InputStream. The "read" call on  InputStream returns
an "Abstract Integer" and value ranges from 0-255. Yup This is what I
needed for reading byte-sequence